Moving test files around and fixing linting issues.
