+++ pages/index.ts
@@ -1,10 +1,11 @@
 import { useRouter } from "next/router";
+import { NextPageContext } from "next";
 
 /**
  * This is the home "/" route. For any requests to this page, redirect
  * to "/library-card/new".
  */
-function RedirectPage({ ctx }) {
+function RedirectPage() {
   const router = useRouter();
   // Make sure it's on the client side.
   if (typeof window !== "undefined") {
@@ -13,13 +14,13 @@ function RedirectPage({ ctx }) {
   }
 }
 
-RedirectPage.getInitialProps = ctx => {
+RedirectPage.getInitialProps = (ctx: NextPageContext) => {
   // Check for ctx.res to make sure we're on the server.
   if (ctx.res) {
     ctx.res.writeHead(302, { Location: "/library-card/new" });
     ctx.res.end();
   }
   return {};
-}
+};
 
 export default RedirectPage;
