+++ pages/library-card/new/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ApplicationContainer from "../../../src/components/ApplicationContainer";
 import { ParamsContextProvider } from "../../../src/context/ParamsContext";
-import LibraryCardForm from "../../../src/components/LibraryCardForm/LibraryCardForm";
+import LibraryCardForm from "../../../src/components/LibraryCardForm";
 
 function HomePage({ query }) {
   return (
