+++ src/components/__tests__/AgeForm.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render, screen, fireEvent, act } from "@testing-library/react";
 import { axe, toHaveNoViolations } from "jest-axe";
 import "@testing-library/jest-dom/extend-expect";
-import AgeForm from "../src/components/AgeForm";
+import AgeForm from "../AgeForm";
 
 expect.extend(toHaveNoViolations);
