+++ src/components/__tests__/ApplicationContainer.test.tsx
@@ -4,7 +4,7 @@ import { render, screen } from "@testing-library/react";
 import { axe, toHaveNoViolations } from "jest-axe";
 import "@testing-library/jest-dom/extend-expect";
 
-import ApplicationContainer from "../src/components/ApplicationContainer";
+import ApplicationContainer from "../ApplicationContainer";
 
 expect.extend(toHaveNoViolations);
